testDAU(): add bg.noise to handle degenerated cases
In degenerated cases (i.e. the column (experiment and the background)
contains single AA type) the diff and SD are 0 for all the AA types in the column.
To avoid this generation, bg.noise > 0 should be specified.
The resulting background AA frequencies are then the linear
combination of the original frequencies with weight (1-bg.noise) and
Dirichlet(1)-distributed random frequencies with weight bg.noise.